fix: handleEdit
